video: Migrate exynos display options to Kconfig
Following how it's done for the majority of drivers, add a new
VIDEO_EXYNOS option and Kconfig file under drivers/video/exynos and list
the current options there.

Cc: Anatolij Gustschin <agust@denx.de>
Cc: Jaehoon Chung <jh80.chung@samsung.com>
Cc: Minkyu Kang <mk7.kang@samsung.com>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
It would be good to have help options here, but I don't know the
underlying parts, so what to add would be appreciated, or done as a
follow-up.  On a related note, the drivers themselves should be under
something in the top-level MAINTAINERS file as they are not currently.