Convert CONFIG_ENV_MIN_ENTRIES et al to Kconfig
This converts the following to Kconfig:
   CONFIG_ENV_MIN_ENTRIES
   CONFIG_ENV_MAX_ENTRIES

Cc: Michal Simek <michal.simek@amd.com>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
I'm not 100% sure that what was being done in xilinx_zynqmp.h was
working before this change.  If we _need_ to tweak this value down in
SPL we need to introduce CONFIG_SPL_ENV_{MIN,MAX}_ENTRIES and use
CONFIG_VAL() in the code.