mx*sabresd: Reference CONFIG_SYS_AUXCORE_BOOTDATA value directly
As this is used in the environment, reference it directly rather than as
a CONFIG value.

Cc: Fabio Estevam <festevam@gmail.com>
Cc: Adrian Alonso <adrian.alonso@nxp.com>
Signed-off-by: Tom Rini <trini@konsulko.com>
---
An alternative would be to move to text file environments (which would
be good to do regardless) and control this via an environmental variable
if desired.  Or introduce a new CONFIG symbol in Kconfig.