global: Add a select for ARCH_MISC_INIT where used
When we have an arch_misc_init function, we can select it to ensure it's
used.

In that there is a functional change here, it's that on i.MX6 it's
always been possible to populate "serial#" with something meaningful,
but not every platform was taking the hook, and now it is.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Cc: Stefano Babic <sbabic@denx.de>
Cc: Fabio Estevam <festevam@gmail.com>
Cc: NXP i.MX U-Boot Team <uboot-imx@nxp.com>

Changes in v2:
- Always select ARCH_MISC_INIT on IMX8M as dealing with FSL_CAAM or not
  needs further SoC-specific clean up work.