bootstd: Enable BOOTSTD_DEFAULTS by default
This is needed to enable the boot command used to start standard boot.
Enable it by default. This brings in quite a few features, mostly in
common with DISTRO_DEFAULTS

Disable this option for boards which don't have enough space.

Disable BOOTSTD_DEFAULTS for smartweb and some sama5 boards since they
are too close to their limit.

Signed-off-by: Simon Glass <sjg@chromium.org>
[trini: Reword slightly]
Signed-off-by: Tom Rini <trini@konsulko.com>