test/py: Fix two problems with setup_image
There are two issues with this function currently. One is that while we
can be passed an image size to use, we always called qemu-img with 20M
as the size. Fix this by using the size parameter. The second issue is
that reserving 2MiB at the end is not enough room to add a GPT (and the
backup GPT that goes on the end) so reserve 4MiB off the end and make
all of the existing callers use 16M rather than 18M.

Signed-off-by: Tom Rini <trini@konsulko.com>
---
Changes in v2:
- None

Cc: Simon Glass <sjg@chromium.org>